HS-95-06: meetings through the desk — record, review, and real intelligence in-world
LiveCore and HistoryCore carry the meeting room and the meeting memory as
host-agnostic cores; flat routes are thin wrappers; the desk hosts both
as windows. Record (chip and orb) starts the hub recorder AND opens the
live window — recording never leaves the desk; Stop inside the window
settles chip, orb, and window (one recorder truth). 'Return to saved
Meeting' opens the pull-out through the shell's openPrimitive
(refresh-then-open, matching the ?open= arrival); 'Review meeting' and
'Record follow-up' open scoped windows. HistoryCore takes scope by prop
(the wrapper still decodes ?room/?meeting for deep links).

The proof is PHYSICAL (evidence-story-06.md + assets/): the hub records
the real microphone (Phase 73 lock), so the walk plays a say-generated
meeting through the speakers at managed volume — Whisper transcribes it
and the .43 llama.cpp titles the meeting and writes the summary; the
review window wears 'Intelligence ready'. Run twice live. Named honestly:
the detail keeps its Dialog grammar inside the window (Article VII drift
for the docs pass), and the golden-43 deck's intel keys were shadowed by
DB-backed settings on this staged run (enabled through the product's own
settings route; rig observation for BACKLOG at closeout). Suite 258/258;
the local leg walks record→stop→pull-out→review→flat routes.

PMO-Story: HS-95-06
PMO-Contract-Digest: sha256:628470a5f788215e06786966ddde48c47de4bb8471d553d562f97f01d89ff047
